+++ src/server.jsx
@@ -28,6 +28,7 @@ import {
   persistAuthToken,
   normalizeLanguageName,
 } from '@plone/volto/helpers';
+import { changeLanguage } from '@plone/volto/actions';
 
 import userSession from '@plone/volto/reducers/userSession/userSession';
 
@@ -195,6 +196,22 @@ server.get('/*', (req, res) => {
 
   loadOnServer({ store, location, routes, api })
     .then(() => {
+      // The content info is in the store at this point thanks to the asynconnect
+      // features, then we can force the current language info into the store when
+      // coming from an SSR request
+      const contentLang =
+        store.getState().content.data?.language?.token ||
+        config.settings.defaultLanguage;
+
+      const cookie_lang =
+        req.universalCookies.get('I18N_LANGUAGE') ||
+        config.settings.defaultLanguage ||
+        req.headers['accept-language'];
+
+      if (cookie_lang !== contentLang) {
+        store.dispatch(changeLanguage(contentLang, locales[contentLang], req));
+      }
+
       const context = {};
       resetServerContext();
       const markup = renderToString(