+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Breaking
 
+- Deprecating `lang` cookie in favor of Plone official one `I18N_LANGUAGE` @sneridagh
+
 ### Feature
 
 - Refactor language syncronizer. Remove it from the React tree, integrate it into the Api Redux middleware @sneridagh